+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 
 ## Bug Fixes
 - [KOGITO-5512](https://issues.redhat.com/browse/KOGITO-5512) - controller-gen doesn't generate CRD's
+- [KOGITO-4470](https://issues.redhat.com/browse/KOGITO-4470) - ConfigMap required ownerReference as controller
 - [KOGITO-5573](https://issues.redhat.com/browse/KOGITO-5573) - Kogito Operator Data-Index Unable to connect to Infinispan Instance
 
 ## Known Issues